requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-bitfield==2.2.0
django-filter==23.5
django-geo==0.8
django-js-asset==2.2.0
django-mptt==0.15.0
django-rest-fuzzysearch==0.5.1
django-separatedvaluesfield==0.6.0
django-timezone-field==6.1.0
django-uuidfield==0.5.0
django-watson==1.6.3
djangorestframework==3.14.0
idna==3.6
lxml==4.9.4
Markdown==3.5.2
psycopg2==2.9.9
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
South==1.0.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
unicodecsv==0.14.1
urllib3==2.1.0
